Move out DNS updater to rpc package
